% Generated by build script
\name{chron_search}
\alias{chron_search}
\title{Search historic newspaper pages on Chronicling America}
\usage{
chron_search()
}
\arguments{
\item{text}{Search text}
\item{page}{Page number of results (default 1, 40 results per page)}
}
\value{tibble: title, date, newspaper, id, image_url, url}
\description{Search historic newspaper pages on Chronicling America}
