% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{loc_collections}
\alias{loc_collections}
\title{List Library of Congress collections}
\usage{
loc_collections()
}
\value{
tibble: id, title, description, url, count
}
\description{
List Library of Congress collections
}
