% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{loc_item}
\alias{loc_item}
\title{Fetch a specific LOC item by URL path}
\usage{
loc_item(url)
}
\arguments{
\item{url}{Full LOC item URL (e.g. "https://www.loc.gov/item/2002717998/")}
}
\value{
tibble: single row with item details
}
\description{
Fetch a specific LOC item by URL path
}
