% Generated by build script
\name{loc_search}
\alias{loc_search}
\title{Search Library of Congress}
\usage{
loc_search()
}
\arguments{
\item{query}{Search term}
\item{collection}{Collection (optional, e.g. 'newspapers', 'photos')}
}
\value{tibble of search results}
\description{Search Library of Congress}
